move mathimg logic to makerhelper. avoid making images multiply
